pyproject.toml
[tool.poetry]
name = "authapp"
version = "0.1.0"
description = ""
authors = ["Voronwe Sul <mail@voronwe.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
requests = "^2.32.3"
python-telegram-bot = "^21.9"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
